requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-decouple==2.1
django-timezone-field==4.2.3
idna==3.3
jmespath==1.0.0
kombu==5.2.4
lxml==4.8.0
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.29
PyJWT==2.3.0
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tweepy==4.8.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
